backend/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.4
Brotli==1.1.0
bs4==0.0.2
cffi==1.17.1
cssselect2==0.8.0
Django==5.2
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
fonttools==4.57.0
pillow==11.2.1
psycopg2-binary==2.9.9
pycparser==2.22
pydyf==0.11.0
PyJWT==2.10.1
pyphen==0.17.2
soupsieve==2.6
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.13.2
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1
